moved cli and zenity frontend to their own files
